14236 signed math leads getelfshdr astray 14242 programs that lack PT_PHDR are not properly loaded Portions contributed by: Bryan Cantrill <bryan@joyent.com> Portions contributed by: Andy Fiddaman <illumos@fiddaman.net> Reviewed by: Robert Mustacchi <rm+illumos@fingolfin.org> Reviewed by: Patrick Mooney <pmooney@pfmooney.com> Reviewed by: Andy Fiddaman <illumos@fiddaman.net> Approved by: Rich Lowe <richlowe@richlowe.net>

+++ usr/src/uts/common/brand/sn1/sn1_brand.c
@@ -21,6 +21,7 @@
 
 /*
  * Copyright (c) 2006, 2010, Oracle and/or its affiliates. All rights reserved.
+ * Copyright 2019 Joyent, Inc.
  */
 
 #include <sys/errno.h>
@@ -57,7 +58,7 @@ void	sn1_forklwp(klwp_t *, klwp_t *);
 void	sn1_freelwp(klwp_t *);
 void	sn1_lwpexit(klwp_t *);
 int	sn1_elfexec(vnode_t *, execa_t *, uarg_t *, intpdata_t *, int,
-	long *, int, caddr_t, cred_t *, int);
+	size_t *, int, caddr_t, cred_t *, int);
 
 /* sn1 brand */
 struct brand_ops sn1_brops = {
@@ -220,8 +221,8 @@ sn1_init_brand_data(zone_t *zone)
 
 int
 sn1_elfexec(vnode_t *vp, execa_t *uap, uarg_t *args, intpdata_t *idatap,
-	int level, long *execsz, int setid, caddr_t exec_file, cred_t *cred,
-	int brand_action)
+    int level, size_t *execsz, int setid, caddr_t exec_file, cred_t *cred,
+    int brand_action)
 {
 	return (brand_solaris_elfexec(vp, uap, args, idatap, level, execsz,
 	    setid, exec_file, cred, brand_action, &sn1_brand, SN1_BRANDNAME,
